fix(p2p): drop partial types
